reorg code
* Remove common module and move bin, key, record, result_code modules to
  top-level aerospike module
* Create new operations, and query modules.
* Rename command module to commands
* Fully specify all paths in macros.
