+++ src/lib.rs
@@ -107,21 +107,36 @@ pub use net::Host;
 pub use value::Value;
 pub use policy::{Policy, ClientPolicy, ReadPolicy, WritePolicy, Priority, ConsistencyLevel,
                  CommitLevel, RecordExistsAction, GenerationPolicy, ScanPolicy, QueryPolicy};
-pub use common::{Key, Bin, Operation, UDFLang, Recordset, Statement, Filter, IndexType,
-                 CollectionIndexType, ParticleType, MapPolicy, MapReturnType, ResultCode};
+pub use key::Key;
+pub use bin::Bin;
+pub use record::Record;
+pub use query::{Statement, Filter, UDFLang, Recordset};
+pub use result_code::ResultCode;
+pub use user::User;
+pub use commands::{IndexType, CollectionIndexType, ParticleType};
+pub use operations::{Operation, MapPolicy, MapReturnType};
 
 #[macro_use]
 pub mod errors;
 
-mod command;
+mod commands;
 mod msgpack;
 mod cluster;
 
-pub mod common;
-pub mod value;
-pub mod policy;
-pub mod net;
+#[macro_use]
+pub mod bin;
 pub mod client;
+#[macro_use]
+pub mod key;
+pub mod net;
+pub mod operations;
+pub mod policy;
+pub mod query;
+pub mod record;
+pub mod result_code;
+pub mod user;
+#[macro_use]
+pub mod value;
 
 #[cfg(test)]
 extern crate hex;
